[chore] asset 파일 분리
