+++ src/assets/Chevron/index.tsx
@@ -1,0 +1,13 @@
+import LeftChevron from './chevron_left.svg';
+import RightChevron from './chevron_right.svg';
+import UpChevron from './chevron_up.svg';
+import VerticalArrows from './verticalArrows.svg';
+import DownChevron from './chevron_down.svg';
+
+export const CHEVRON = {
+  down: <DownChevron />,
+  verticalArrows: <VerticalArrows />,
+  left: <LeftChevron />,
+  right: <RightChevron />,
+  up: <UpChevron />,
+};
