+++ src/assets/Post/index.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import Declation from './declation.svg';
+import Edit from './edit.svg';
+import Copy from './copy.svg';
+
+export const POST_SVG = {
+  declation: <Declation />,
+  edit: <Edit />,
+  copy: <Copy />,
+};
