+++ src/components/CreateCopy/CopyListItem/index.tsx
@@ -1,6 +1,9 @@
 import React, { useState } from 'react';
 import { useRecoilState } from 'recoil';
-import { PIN, POST_SVG, SVG } from '../../../assets';
+import { SVG } from '../../../assets';
+import { PIN } from '../../../assets/Like';
+import { POST_SVG } from '../../../assets/Post';
+
 import { copyListState } from '../../../store/copyListState';
 import { CopyListType } from '../../../types/copy';
 import Button from '../../common/Button';
